+++ src/components/spa/main.tsx
@@ -7,6 +7,7 @@ import { CTF } from "components/ctf/ctf";
 import { ProcessingTitle } from "components/visualisation/processingTitle";
 import { BaseProcessingJobProps } from "schema/interfaces";
 import { recipeTagMap } from "utils/config/parse";
+import { RefinementStep } from "./refine";
 
 // This refinement step blacklist should be TEMPORARY and will be removed when a proper data view exists
 const checkRecipe = (target: string, procJob: BaseProcessingJobProps["procJob"]) =>
@@ -19,15 +20,21 @@ const SPA = ({ autoProc, procJob, status, active }: BaseProcessingJobProps) => {
 
   const toDisplay = useMemo(
     () =>
-      ["em-spa-preprocess", "em-spa-class2d", "em-spa-class3d"].map((target) =>
+      ["em-spa-preprocess", "em-spa-class2d", "em-spa-class3d", "em-spa-refine"].map((target) =>
         checkRecipe(target, procJob)
       ),
     [procJob]
   );
 
+  // TODO: remove all references of beta in processing title once refinement row is out of beta
   return (
     <AccordionItem>
-      <ProcessingTitle autoProc={autoProc} procJob={procJob} status={status} />
+      <ProcessingTitle
+        autoProc={autoProc}
+        procJob={procJob}
+        status={status}
+        isBeta={toDisplay[3]}
+      />
       <AccordionPanel p={4} bg='diamond.75'>
         {autoProc ? (
           <>
@@ -58,6 +65,7 @@ const SPA = ({ autoProc, procJob, status, active }: BaseProcessingJobProps) => {
                 {toDisplay[2] && (
                   <Classification autoProcId={autoProc.autoProcProgramId} type='3d' />
                 )}
+                {toDisplay[3] && <RefinementStep autoProcId={autoProc.autoProcProgramId} />}
               </Grid>
             )}
           </>